Fix overlapping_rect's screen coordinate calculation in find bar
The find bar is supposed to move when it is intersected with the selected text rectangle.
To decide the intersection, the selected text rect needs to be translated to the screen
coordinate first.

Currently the webcontents_rect is the web page rect in its *parent's* coordinate which is
wrong for calculating the translation for avoid_overlapping_rect.

The patch fixes the translation by using webcontents_rect's screen coordinate. 


BUG=
TEST= FindInPageControllerTest.FindMovesWhenObscuring in browser_tests pass for touch build


Review URL: http://codereview.chromium.org/7083002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@87059 0039d316-1c4b-4281-b951-d872f2087c98