Random #14 trivial provided by wikipedia:
The number of days in a fortnight.
In traditional British units of weight, the number of pounds in a stone.
A number 'encoded' in much of the music of Johann Sebastian Bach.[citation needed] Bach may have considered this number a sort of signature, since given A = 1, B = 2, C = 3, etc., then B + A + C + H = 14. (See also 41)
The number of points outlined by president Woodrow Wilson for reconstructing a new Europe following World War I, see Fourteen Points.
The section that you go to when you die in the Grailquest books.
The number of legs on a woodlouse, as well as on Hallucigenia.
A common designation for the thirteenth floor in many buildings for superstitious reasons
The number of points in a proposed republican constitution of the United Kingdom[citation needed]
The number of lines in a sonnet.
The Number 14 airship by Alberto Santos Dumont that was used to test the aerodynamics of his 14-bis airplane.
The number of the French department Calvados
A Storage server manufactured by IBM. It goes by name of "XIV" and is pronounced as the separate letters "X", "I", "V".

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@87559 0039d316-1c4b-4281-b951-d872f2087c98