Revert 88137 - Add code to calculate the dominant color for a favicon.
Currently we calculate the dominant/representative color only for those favicons we need it for (i.e. the ones shown on the NTP). We don't do any caching either in memory or in the favicon db but that can be tacked on later if deemed suitable.

Code in color_analysis.* authored by dtrainor

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/7099001

TBR=estade@chromium.org
Review URL: http://codereview.chromium.org/7125010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@88139 0039d316-1c4b-4281-b951-d872f2087c98