Fix crash when you uninstall a permissions-upgrade disabled extension.
When an extension is disabled because a newer version requires higher
permissions, if you then uninstall the extension while the infobar warning
is still up, you'll get a crash if you switch to the tab that had the infobar.
This patch makes the infobar notice the extension uninstall and remove itself.

BUG=84826
TEST=As detailed in description above (and outlined in bug report too).


Review URL: http://codereview.chromium.org/7003068

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@88458 0039d316-1c4b-4281-b951-d872f2087c98