Generic interface-based unit testing for BaseTabStrip
This provides some setup and infrastructure so that you can write tests for BaseTabStrip, but perform those tests against all implementations of BaseTabStrip (i.e.: TabStrip, SideTabStrip, TouchTabStrip, etc.)

I wrote a couple of basic tests just to get the ball rolling.  I also made a little fix in TouchTabStrip to get my example AddTabAt test to pass with TouchTabStrip.

I made use of a parameterized include file (see notes in source).  This is so that each individual test can still use the gtest macros to get its own isolated test (as opposed to writing a TestEverything() function.)  The idea is to leverage as much as possible of the services that gtest provides, and still work around the problem that gtest doesn't allow you to declare templatized tests, and the test class names must be unique.  I tried my hardest to get this to conform to Chromium style, but it does involve defining a macro before including the header file.  Hopefully this is tolerable.  The advantage, again, is that you get to retain all the advantages of individual gtest test cases without changing anything in gtest itself, and you can still reuse test code against other implementation classes with minimal code duplication.

This also provides a starting point for having unit tests for the TabStrip, SideTabStrip and TouchTabStrip.

The tests that I included are ultra-simple, almost tautological.  They are just supposed to illustrate how the tests should be written.  It should be expected that more meaningful tests are coming in the future.

BUG=None
TEST=unit_tests --gtest_filter=*BaseTabStrip*

Review URL: http://codereview.chromium.org/7080030

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@88991 0039d316-1c4b-4281-b951-d872f2087c98