Autofill DataDrivenHeuristics tests tuning
This change eliminates parts of html test suite not related to <form> elements in an attempt to speed up the tests.

Before:
  real	0m58.920s
  user	1m4.858s
  sys	0m15.006s

After:
  real	0m46.897s
  user	0m50.286s
  sys	0m12.361s

A subsequent patch will attempt to re-enable these tests on Windows in the hopes that we're under the wire now with respect to timeouts.

BUG=85276
TEST=FormStructureBrowserTest.*

Review URL: http://codereview.chromium.org/7190012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@89396 0039d316-1c4b-4281-b951-d872f2087c98