Chrome side of the infobar experiment
When the browser is asked for the plugin policy, it checks for an override that is in the registry, this
override has been set by setup.exe (if and only if) the user has been drafted for the experiment.

See http://codereview.chromium.org/7148004 for the experiment selection part.

TEST= see bug
BUG=85988
Review URL: http://codereview.chromium.org/7233017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@90461 0039d316-1c4b-4281-b951-d872f2087c98