Expose CompositorGL and TextureGL.
RWHVV_touch will soon use CompositorGL and TextureGL. It will subclass TextureGL to wrap image data coming from the GPU process. It will need to access CompositorGL in order to make the GL context associated with the compositor current when making instances of the subclass of TextureGL.

In addition, the subclass of TextureGL will use a slightly different fragment shader. This CL also includes a refactor of how shaders are produced to share commonality.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/7235003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@90506 0039d316-1c4b-4281-b951-d872f2087c98