Fixup error message shown when Chromoting is not enabled properly.
Previously, we did a version check first, so if the plugin didn't load we
would display an (inappropriate) "version mis-match, please upgrade" message.
This was very frustrating since it often happened immediately after
installation, when the user was trying to use Remoting for the first time.

This change checks the plugin status first and (if there's a problem) directs
the user to verify that the 'Remoting' and 'P2P API' flags are enabled.

This partially addresses bug 87282, but we still need to make this look
prettier.

BUG=87282
TEST=manual
Review URL: http://codereview.chromium.org/7263008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@90626 0039d316-1c4b-4281-b951-d872f2087c98