Convert RenderViewContextMenu to MenuItemView.
This CL is part of general GTK removal for ChromiumOS.  Menu2 uses GTK on linux so we are replacing it with MenuItemView.  Chrome Frame currently passes the context menu between processes by using the HMENU.  Because MenuItemView does not use HMENU, we need to use another mechanism.

This CL creates a ContextMenuModel struct that is serialized into an automation message for Chrome Frame.  ContextMenuModel contains the context menu definition in-band replacing the out-of-band HMENU.

BUG=chromium-os:13887
TEST=none


Review URL: http://codereview.chromium.org/7167002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@92182 0039d316-1c4b-4281-b951-d872f2087c98