Trivial fix for IDL to H delta
ppp_instance has a block of injected code:

typedef PPP_Instance PPP_Instance_1_0;

This was not updated to _1_0 from _0_5 when the version
number was bumped.  Since #include blocks are injected
as is, instead of generated, changes that are version
dependent will create problems later.

This CL fixes the IDL so that it generates the correct
thing, however, a better fix would be to either always
generate type typedef automatically, or to not rely on
the typedef in this way.  This particular issue is more
confusing in that it is the only instance of this
requirement.


TEST= trybots
BUG= none

R= brettw@chromium.org
Review URL: http://codereview.chromium.org/7472015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@93291 0039d316-1c4b-4281-b951-d872f2087c98