GTK: Final experiment with Unity Global Menus before I give up on icons.
If you previously had issues with bad lag on Ubuntu Natty, please try this build with the command line flag "--enable-icons-in-global-history-menu". If this both lets icons be in the history menu and makes chrome startup instant, we can roll this out by default.

Revert my patch that just removed icons all together and insert some defensive
coding to check the output of the favicon database. If an icon is larger than
128 pixels, assume corruption and fallback to the default icon. If it's larger
than 16x16, resize it down to 16x16.

This reverts commit 924f4b128794a3b0d838e06517709de8ba6ecaff (r91615).

BUG=86715
TEST=none


Review URL: http://codereview.chromium.org/7486003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@93575 0039d316-1c4b-4281-b951-d872f2087c98