Update pkg-dmg to include *.so files, normally excluded by rsync -C (--cvs-exclude).
Since r94390, the ffmpeg code is in a loadable_module instead of a
shared_library, and its name on disk has changed from libffmpegsumo.dylib to
ffmpegsumo.so. The disk image packager, pkg-dmg, uses rsync with the -C flag,
which excludes *.so files. This change makes pkg-dmg tell rsync to explicitly
include *.so files.

This is similar to r94388.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/7566007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@95249 0039d316-1c4b-4281-b951-d872f2087c98