Allow sync integration tests to operate on multiple datatypes: Typed Urls
The sync integration tests currently use a class hierarchy where the test classes for each datatype are subclasses of LiveSyncTest. While this design worked in the past, it allows tests to work with only one datatype at a time, and therefore doesn't allow us to test the interplay between datatypes.

This patch is another in the series of patches that will move away from an inheritance model to one where test cases can operate on more than one datatype. It updates the Typed Urls datatype to the new model, and contains the following changes:

- LiveTypedUrlsSyncTest is no longer a class that inherits from LiveSyncTest, but is now a namespace called typed_urls_helper, and contains a bunch of methods that perform various operations related to Typed Urls.
- SingleClientLiveTypedUrlsSyncTest is renamed to SingleClientTypedUrlsSyncTest, and is a subclass of LiveSyncTest. It uses the methods in namespace typed_urls_helper by including its header file and attaching itself to the helper.

In addition, bookmarks, preferences and autofill helpers and the SyncDatatypeHelper were originally implemented as static classes. Based on the code review feedback in this patch, they have been modified to the "functions-in-a-namespace" model similar to typed urls. This patch consists of those changes too.

BUG=88510
TEST=sync_integration_tests, sync_performance_tests

Review URL: http://codereview.chromium.org/7461109

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@95862 0039d316-1c4b-4281-b951-d872f2087c98